pyproject.toml
[tool.poetry]
name = "writer-framework-default"
version = "0.1.0"
description = ""
authors = ["Writer Example <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
langgraph = "^0.2.61"
langchain-community = "^0.3.14"
langchain-anthropic = "^0.3.1"
beautifulsoup4 = "^4.12.3"
langchain = "^0.3.14"
langchain-core = "^0.3.29"
writer = "^0.8.2"
python-dotenv = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
